requirements.txt

pillow==11.1.0
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3

